requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1
stripe==7.8.0
requests==2.31.0
pydantic==2.5.0
pydantic-settings==2.1.0
pathlib==1.0.1
gunicorn==21.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
PyPDF2==3.0.1
fpdf==1.7.2
PyMuPDF>=1.23.0
pdf2image>=1.17.0
Pillow>=10.0.0
itsdangerous==2.1.2
numpy



